refactor(crm): Rename crm_meditor file to crm_mediator
